fix LCG
