+++ generators/src/linear_congruential_generator.cpp
@@ -1,0 +1,1 @@
+#include "linear_congruential_generator.hpp"
